feat!: flatten module.
