+++ go.mod
@@ -1,25 +1,10 @@
-module github.com/matthewhartstonge/hasher
+module github.com/matthewhartstonge/hasher/v4
 
-go 1.13
+go 1.16
 
 require (
-	github.com/Sirupsen/logrus v1.0.5 // indirect
-	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/go-errors/errors v1.1.1 // indirect
-	github.com/golang/mock v1.5.0 // indirect
-	github.com/lhecker/argon2 v1.0.0
-	github.com/oleiade/reflections v1.0.1 // indirect
-	github.com/onsi/ginkgo v1.16.1 // indirect
-	github.com/onsi/gomega v1.11.0 // indirect
-	github.com/ory-am/common v0.4.0 // indirect
-	github.com/ory/common v0.4.0 // indirect
-	github.com/ory/fosite v0.8.0
-	github.com/pborman/uuid v1.2.0
-	github.com/pkg/errors v0.8.0
-	github.com/sirupsen/logrus v1.8.1 // indirect
-	github.com/square/go-jose v2.5.1+incompatible // indirect
-	github.com/stretchr/testify v1.6.1
-	gopkg.in/airbrake/gobrake.v2 v2.0.9 // indirect
-	gopkg.in/gemnasium/logrus-airbrake-hook.v2 v2.1.2 // indirect
+	github.com/matthewhartstonge/argon2 v0.2.1
+	github.com/ory/fosite v0.42.2
+	github.com/pkg/errors v0.9.1
+	github.com/stretchr/testify v1.7.0
 )
