+++ hash_argon2_test.go
@@ -1,13 +1,27 @@
 package hasher_test
 
 import (
+	// Standard Library Imports
+	"context"
 	"testing"
 
-	"github.com/lhecker/argon2"
-	"github.com/matthewhartstonge/hasher"
-	"github.com/pborman/uuid"
+	// External Imports
+	"github.com/matthewhartstonge/argon2"
+	"github.com/ory/fosite"
 	"github.com/stretchr/testify/assert"
+
+	// Internal Imports
+	"github.com/matthewhartstonge/hasher/v4"
 )
+
+// TestArgon2_ImplementsHasher ensures that Argon2 implements the fosite.Hasher
+// interface.
+func TestArgon2_ImplementsHasher(t *testing.T) {
+	v := &hasher.Argon2{}
+	var i interface{} = v
+	_, ok := i.(fosite.Hasher)
+	assert.Equal(t, true, ok)
+}
 
 // TestArgon2Hash ensures that a hash is returned from the Argon2 Hasher
 func TestArgon2Hash(t *testing.T) {
@@ -15,22 +29,9 @@ func TestArgon2Hash(t *testing.T) {
 		Config: argon2.DefaultConfig(),
 	}
 	password := []byte("foo")
-	hash, err := h.Hash(password)
+	hash, err := h.Hash(context.Background(), password)
 	assert.Nil(t, err)
 	assert.NotNil(t, hash)
-	assert.NotEqual(t, hash, password)
-}
-
-// TestArgon2HashLibErr purposely causes the underlying argon2 lib to error to ensure it is reported up the stack
-func TestArgon2HashLibErr(t *testing.T) {
-	h := &hasher.Argon2{
-		Config: argon2.DefaultConfig(),
-	}
-	h.Config.MemoryCost = 1
-	password := []byte("foo")
-	hash, err := h.Hash(password)
-	assert.Empty(t, hash)
-	assert.NotNil(t, err)
 	assert.NotEqual(t, hash, password)
 }
 
@@ -40,10 +41,10 @@ func TestArgon2CompareEquals(t *testing.T) {
 		Config: argon2.DefaultConfig(),
 	}
 	password := []byte("foo")
-	hash, err := h.Hash(password)
+	hash, err := h.Hash(context.Background(), password)
 	assert.Nil(t, err)
 	assert.NotNil(t, hash)
-	err = h.Compare(hash, password)
+	err = h.Compare(context.Background(), hash, password)
 	assert.Nil(t, err)
 }
 
@@ -53,23 +54,9 @@ func TestArgon2CompareDifferent(t *testing.T) {
 		Config: argon2.DefaultConfig(),
 	}
 	password := []byte("foo")
-	hash, err := h.Hash(password)
+	hash, err := h.Hash(context.Background(), password)
 	assert.Nil(t, err)
 	assert.NotNil(t, hash)
-	err = h.Compare(hash, []byte(uuid.NewRandom()))
-	assert.NotNil(t, err)
-}
-
-// TestArgon2HashLibErr purposely causes the underlying argon2 lib to error to ensure it is reported up the stack
-func TestArgon2CompareLibErr(t *testing.T) {
-	h := &hasher.Argon2{
-		Config: argon2.DefaultConfig(),
-	}
-	h.Config.MemoryCost = 1
-	password := []byte("foo")
-	hash, err := h.Hash(password)
-	assert.Empty(t, hash)
-	assert.NotNil(t, err)
-	err = h.Compare(hash, []byte(uuid.NewRandom()))
+	err = h.Compare(context.Background(), hash, []byte("911650fc-df29-4622-8c6f-f43cbacd1ece"))
 	assert.NotNil(t, err)
 }
