--- v2/go.mod
@@ -1,18 +1,0 @@
-module github.com/matthewhartstonge/hasher/v2
-
-go 1.14
-
-require (
-	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/golang/mock v1.5.0 // indirect
-	github.com/matthewhartstonge/argon2 v0.1.1
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
-	github.com/oleiade/reflections v1.0.1 // indirect
-	github.com/ory/fosite v0.11.4
-	github.com/pborman/uuid v1.2.1 // indirect
-	github.com/pkg/errors v0.8.0
-	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20210421170649-83a5a9bb288b // indirect
-)
