--- v3/go.mod
@@ -1,10 +1,0 @@
-module github.com/matthewhartstonge/hasher/v3
-
-go 1.14
-
-require (
-	github.com/matthewhartstonge/argon2 v0.1.4
-	github.com/ory/fosite v0.25.0
-	github.com/pkg/errors v0.8.0
-	github.com/stretchr/testify v1.2.2
-)
