--- v4/go.mod
@@ -1,10 +1,0 @@
-module github.com/matthewhartstonge/hasher/v4
-
-go 1.16
-
-require (
-	github.com/matthewhartstonge/argon2 v0.2.1
-	github.com/ory/fosite v0.42.2
-	github.com/pkg/errors v0.9.1
-	github.com/stretchr/testify v1.7.0
-)
